pyproject.toml
[project]
name = "moya-for-research"
version = "0.1.0"
description = "Multi-agent research paper analysis system using MOYA framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "moya-ai",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.6",
    "pypdf2>=3.0.1",
    "sqlalchemy>=2.0.44",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "tenacity>=8.2.3",
]

[project.scripts]
moya-research = "moya_for_research.cli:main"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
moya-ai = { git = "https://github.com/montycloud/moya.git" }

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "ruff>=0.14.5",
]
